refactor: 리포지토리 쿼리 리팩터링
