+++ src/main/java/com/example/mypetlife/entity/article/Article.java
@@ -31,7 +31,7 @@ public class Article {
     @Column(name = "post_date")
     private LocalDateTime postDate;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
