+++ src/main/java/com/example/mypetlife/entity/article/ArticleImage.java
@@ -18,7 +18,7 @@ public class ArticleImage {
     @Column(name = "image_url")
     private String imageUrl;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "article_id")
     private Article article;
