+++ src/main/java/com/example/mypetlife/entity/article/ArticleTag.java
@@ -16,11 +16,11 @@ public class ArticleTag {
     @Column(name = "article_tag_id")
     private Long id;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "article_id")
     Article article;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "tag_id")
     Tag tag;
