+++ src/main/java/com/example/mypetlife/entity/article/LikeArticle.java
@@ -15,11 +15,11 @@ public class LikeArticle {
     @Column(name = "like_article_id")
     private Long id;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "article_id")
     private Article article;
