+++ src/main/java/com/example/mypetlife/service/community/TagService.java
@@ -4,36 +4,42 @@ import com.example.mypetlife.entity.article.Tag;
 import com.example.mypetlife.exception.CustomException;
 import com.example.mypetlife.exception.ErrorCode;
 import com.example.mypetlife.repository.community.TagRepository;
-import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
 
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class TagService {
 
     private final TagRepository tagRepository;
-    private final EntityManager em;
 
-    public boolean isExistInDb(String tagName) {
+    /*
+     * DB에 없는 새로운 태그인지 확인
+     */
+    public boolean isNewTag(String tagName) {
 
-        Optional<Tag> optionalTag = tagRepository.findByTagName(tagName);
-        if(optionalTag.isEmpty()) {
-            return false;
-        }
-        return true;
+        return !tagRepository.findByTagName(tagName)
+                .isPresent();
     }
 
+    /*
+     * 태그이름으로 조회
+     */
     public Tag findByTagName(String tagName) {
+
         return tagRepository.findByTagName(tagName)
                 .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_TAG));
     }
 
+    /*
+     * 저장
+     */
+    @Transactional
     public void saveTag(Tag tag) {
+
         tagRepository.save(tag);
     }
